refactor(livechat): move checkRoom and grantUser to the connector
Extract checkRoom and grantUser from the ChatContainer class into the
connector as useStableCallback hooks, passed down as props. The container
keeps its lifecycle but now calls these via props, and drops its direct
store import. Enable the useStableCallback helper.
